refactor: update `LuaJIT` IoC namespace structure and adjust imports accordingly
