+++ src/Extensions/NeoServer.Scripts.LuaJIT/IoC/LuaJITInjection.cs
@@ -5,7 +5,7 @@ using NeoServer.Scripts.LuaJIT.Interfaces;
 using NeoServer.Scripts.LuaJIT.Managers;
 using NeoServer.Scripts.LuaJIT.Services;
 
-namespace NeoServer.Scripts.LuaJIT.IoC.Modules;
+namespace NeoServer.Scripts.LuaJIT.IoC;
 
 public static class LuaJitInjection
 {
